restructure family variant details
